requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
pydantic==2.10.5
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
